requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
oslo.config>=1.4.0
PasteDeploy==1.5.2
oslo.messaging
oslo.db
oslo.log
oslo.i18n
oslo.utils>=1.2.0                       # Apache-2.0
pecan>=0.8
keystonemiddleware>=1.0.0
six>=1.7.0,<=1.9.0
sqlalchemy
stevedore>=1.1.0  						# Apache-2.0
WSME>=0.6
jsonpatch>=1.1
enum34==1.0.4

# watcher Applier
python-novaclient
python-openstackclient
python-neutronclient
python-glanceclient
python-cinderclient
# Decision Engine
python-ceilometerclient